{{cookiecutter.project_slug}}/requirements/base.txt
Django==5.0.8
django-environ==0.11.2
django-extensions==3.2.3
django-model-utils==4.5.1
django-widget-tweaks==1.5.0
ipython==8.26.0
django-safedelete==1.4.0
django-simple-history==3.7.0
